+++ ERControl/src/main/java/er/extensions/components/woextensions/events/WXOutlineRepetition.java
@@ -1,0 +1,23 @@
+/*
+ * WXOutlineRepetition.java
+ * (c) Copyright 2001 Apple Computer, Inc. All rights reserved.
+ * This a modified version.
+ * Original license: http://www.opensource.apple.com/apsl/
+ */
+
+package er.extensions.components.woextensions.events;
+
+import com.webobjects.appserver.WOComponent;
+import com.webobjects.appserver.WOContext;
+
+public class WXOutlineRepetition extends WOComponent {
+
+	public WXOutlineRepetition(WOContext aContext) {
+		super(aContext);
+	}
+
+	@Override
+	public boolean synchronizesVariablesWithBindings() {
+		return false;
+	}
+}